Refactoring Performed Extract Class Refactoring Added Comments
